+++ ackpine-splits/src/main/kotlin/ru/solrudev/ackpine/splits/ZippedApkSplits.kt
@@ -19,7 +19,9 @@ package ru.solrudev.ackpine.splits
 import android.content.Context
 import android.net.Uri
 import android.os.Build
+import android.os.ParcelFileDescriptor
 import androidx.annotation.RequiresApi
+import ru.solrudev.ackpine.helpers.closeWithException
 import ru.solrudev.ackpine.helpers.entries
 import ru.solrudev.ackpine.helpers.toFile
 import java.io.File
@@ -64,12 +66,7 @@ public object ZippedApkSplits {
 				return@closeableSequence
 			}
 			if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-				try {
-					yieldAllUsingFileChannel(applicationContext, uri)
-				} catch (exception: Exception) {
-					exception.printStackTrace()
-					yieldAllUsingZipInputStream(applicationContext, uri)
-				}
+				yieldAllApi26(applicationContext, uri)
 				return@closeableSequence
 			}
 			yieldAllUsingZipInputStream(applicationContext, uri)
@@ -92,24 +89,42 @@ public object ZippedApkSplits {
 	}
 
 	@RequiresApi(Build.VERSION_CODES.O)
-	private suspend inline fun CloseableSequenceScope<Apk>.yieldAllUsingFileChannel(context: Context, uri: Uri) {
-		val fd = context.contentResolver.openFileDescriptor(uri, "r")?.use()
-			?: throw NullPointerException("ParcelFileDescriptor was null: $uri")
-		val fileInputStream = FileInputStream(fd.fileDescriptor).use()
-		val zipFile = org.apache.commons.compress.archivers.zip.ZipFile.builder()
-			.setSeekableByteChannel(fileInputStream.channel)
-			.get()
-			.use()
-		zipFile.entries
-			.asSequence()
-			.filterNot { isClosed }
-			.mapNotNull { zipEntry ->
-				zipFile.getInputStream(zipEntry).use { entryStream ->
-					entryStream.use()
-					Apk.fromZipEntry(uri.toString(), zipEntry, entryStream)
+	private suspend inline fun CloseableSequenceScope<Apk>.yieldAllApi26(context: Context, uri: Uri) {
+		var fd: ParcelFileDescriptor? = null
+		var fileInputStream: FileInputStream? = null
+		val zipFile: org.apache.commons.compress.archivers.zip.ZipFile
+		try {
+			fd = context.contentResolver.openFileDescriptor(uri, "r")?.use()
+				?: throw NullPointerException("ParcelFileDescriptor was null: $uri")
+			fileInputStream = FileInputStream(fd.fileDescriptor).use()
+			zipFile = org.apache.commons.compress.archivers.zip.ZipFile.builder()
+				.setSeekableByteChannel(fileInputStream.channel)
+				.get()
+				.use()
+		} catch (throwable: Throwable) {
+			fd?.closeWithException(throwable)
+			fileInputStream?.closeWithException(throwable)
+			throwable.printStackTrace()
+			yieldAllUsingZipInputStream(context, uri)
+			return
+		}
+		try {
+			zipFile.entries
+				.asSequence()
+				.filterNot { isClosed }
+				.mapNotNull { zipEntry ->
+					zipFile.getInputStream(zipEntry).use { entryStream ->
+						entryStream.use()
+						Apk.fromZipEntry(uri.toString(), zipEntry, entryStream)
+					}
 				}
-			}
-			.forEach { yield(it) }
+				.forEach { yield(it) }
+		} catch (throwable: Throwable) {
+			fd.closeWithException(throwable)
+			fileInputStream.closeWithException(throwable)
+			zipFile.closeWithException(throwable)
+			throw throwable
+		}
 	}
 
 	private suspend inline fun CloseableSequenceScope<Apk>.yieldAllUsingZipInputStream(context: Context, uri: Uri) {